Initial commit for update_title_bar.
update_title_bar automatically updates the terminal window's title,
based on what is running in the terminal.

Many terminal emulators provide this functionality internally (parsing the
text in the terminal window buffer), but many do not. I prefer to use
Terminator, which can be updated, but only with ANSI escape calls. This
small tool will add itself to the local PROMPT_COMMAND.

* It should not cause any issues with pre-existing executables listed,
* but YMMV. If this tool interferes with your existing PROMPT_COMMAND, please
* let me know by submitting an issue via Github.