addpath now supports custom variables.
New feature for addpath that allows you to specify the variable name to modify. It can be specified by prepending the addpath command with "AP_VAR_NAME=[var name]", where "[var name]" is the name of the variable to use.

Also added ability to specify a custom delimiter (AP_VAR_DELIM variable), instead of ':' (This was done in previous commit, but I'm feeeling too lazy to edit the log entry).